requirements.txt
mock
polib
pytest
pytest-mypy
pytest-flake8
pytest-isort
pytest-black
rapidjson
requests

flask
flask-babel
flask-login
flask-sqlalchemy
psycopg2-binary
pyTelegramBotAPI
SQLAlchemy
pyyaml
werkzeug
